requirements.txt
appdirs==1.4.4
audioread==3.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Flask==2.3.2
Flask-WTF==1.1.1
fonttools==4.40.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.1
msgpack==1.0.5
numba==0.57.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
